// src/gen/types/ProductResponse.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type ProductResponse = {
    /**
     * @type string, uuid
    */
    readonly id: string;
    /**
     * @type string
    */
    readonly name: string;
    /**
     * @pattern ^-?\d{0,8}(?:\.\d{0,2})?$
     * @type string, decimal
    */
    readonly unit_price: string;
    /**
     * @type integer
    */
    readonly quantity: number;
    /**
     * @pattern ^-?\d{0,10}(?:\.\d{0,2})?$
     * @type string, decimal
    */
    readonly total_value: string;
    /**
     * @type string, date-time
    */
    readonly created_at: string;
    /**
     * @type string, date-time
    */
    readonly updated_at: string;
};